Separate ReactShallowRenderer from TestUtils
